requirements.txt
pyTelegramBotAPI==4.4.0
python-dotenv==0.19.2
telebot-calendar==1.2
peewee==3.15.2